# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "LoginRequest": {
        "properties": {
          "password": {
            "example": "Admin#2026",
            "maxLength": 200,
            "minLength": 0,
            "type": "string"
          },
          "username": {
            "example": "admin",
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "password",
          "username"
        ],
        "type": "object"
      },
...
      "RefreshTokenRequest": {
        "properties": {
          "refreshToken": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ],
        "type": "object"
      },
...
      "TokenResponse": {
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "expiresInSeconds": {
            "example": 1800,
            "format": "int64",
            "type": "integer"
          },
          "refreshToken": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "type": "object"
      },
...
      "User": {
        "properties": {
          "fullName": {
            "example": "System Administrator",
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "role": {
            "enum": [
              "ADMIN",
              "SALES"
            ],
            "type": "string"
          },
          "username": {
            "example": "admin",
            "type": "string"
          }
        },
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "Access token from POST /api/auth/login",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Abdalrahman Bsharat"
    },
    "description": "REST API for the Triosuite ERP sales-invoice assessment.\n\nMulti-currency invoices with user-supplied exchange rates, tax-inclusive\nand tax-exclusive calculation, server-side automatic numbering, an\napproval and cancellation state machine, and barcode lookup.\n\nAuthenticate with `POST /api/auth/login` (try `admin` / `Admin#2026`),\nthen paste the returned `accessToken` into **Authorize**.\n\nErrors are RFC 9457 problem documents carrying a stable `code`\nproperty; see docs/API.md for the full list.",
    "license": {
      "name": "MIT"
    },
    "title": "Triosuite Invoices API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/auth/login": {
      "post": {
        "description": "Exchanges a username and password for an access token and a refresh token.\n\nRate limited to five failed attempts per minute, counted separately per username\nand per client IP; exceeding either returns 429 with a Retry-After header.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Signed in"
          },
          "400": {
            "description": "Missing username or password"
          },
          "401": {
            "description": "Incorrect username or password"
          },
          "429": {
            "description": "Too many attempts"
          }
        },
        "security": [],
        "summary": "Sign in",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "description": "Revokes a refresh token. Idempotent: an unknown or already revoked token also\nreturns 204, so the endpoint cannot be used to probe which tokens exist.",
        "operationId": "logout",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Token revoked"
          }
        },
        "summary": "Sign out",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "description": "Resolves the caller's access token to an account.",
        "operationId": "me",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "description": "The signed-in account"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "summary": "The signed-in account",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "description": "Exchanges a refresh token for a new pair. The presented token is revoked in the\nsame transaction, so each one works exactly once.",
        "operationId": "refresh",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "New token pair issued"
          },
          "401": {
            "description": "Unknown, revoked or expired refresh token"
          }
        },
        "security": [],
        "summary": "Rotate tokens",
        "tags": [
          "Authentication"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://127.0.0.1:8080"
    }
  ],
...
}